[iOS] Test: 테스트 코드 파일 분리
