+++ iOS/MiniVIBE/MiniVIBETests/LocalRepositoryTests.swift
@@ -52,26 +52,3 @@ class LocalRepositoryTests: XCTestCase {
         XCTAssertTrue(repository.fetchEvents().isEmpty)
     }
 }
-
-class FailPersistenceController: PersistenceController {
-    var persistentContainer: NSPersistentContainer = {
-        let container = NSPersistentContainer(name: "test")
-        return container
-    }()
-    
-    var context: NSManagedObjectContext {
-        return persistentContainer.viewContext
-    }
-    
-    func fetch<T>(request: NSFetchRequest<T>) -> [T] where T : NSManagedObject {
-        return []
-    }
-    
-    func delete<T>(request: NSFetchRequest<T>) -> Bool where T : NSManagedObject {
-        return false
-    }
-    
-    func saveEvent(event: CoreDataEvent) -> Bool {
-        return false
-    }
-}
