+++ iOS/MiniVIBE/MiniVIBETests/TestPersistenceController.swift
@@ -9,22 +9,21 @@ import Foundation
 import CoreData
 @testable import MiniVIBE
 
-class TestPersistenceController: RealPersistenceController {
-  override init() {
-    super.init()
-
-    let persistentStoreDescription = NSPersistentStoreDescription()
-    persistentStoreDescription.type = NSInMemoryStoreType
-
-    let container = NSPersistentContainer(name: RealPersistenceController.modelName, managedObjectModel: RealPersistenceController.model)
-    container.persistentStoreDescriptions = [persistentStoreDescription]
-
-    container.loadPersistentStores { (_, error) in
-      if let error = error as NSError? {
-        fatalError(
-          "Unresolved error \(error), \(error.userInfo)")
-      }
+final class TestPersistenceController: RealPersistenceController {
+    override init() {
+        super.init()
+        
+        let persistentStoreDescription = NSPersistentStoreDescription()
+        persistentStoreDescription.type = NSInMemoryStoreType
+        
+        let container = NSPersistentContainer(name: RealPersistenceController.modelName, managedObjectModel: RealPersistenceController.model)
+        container.persistentStoreDescriptions = [persistentStoreDescription]
+        
+        container.loadPersistentStores { (_, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        }
+        self.persistentContainer = container
     }
-    self.persistentContainer = container
-  }
 }
